+++ .github/workflows/database-bigquery-integration-test.yml
@@ -79,4 +79,3 @@ jobs:
           AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_INTEGRATION_USER1_SECRET_ACCESS_KEY }}
         run: |
           mvn -P run-bigquery-integration -pl legend-engine-xt-relationalStore-bigquery-execution-tests -Dtest="ExternalIntegration*BigQuery" test
-          mvn -P run-bigquery-integration -pl legend-engine-xt-relationalStore-executionPlan-connection-authentication-default -Dtest="ExternalIntegration*BigQuery*" test
