+++ legend-engine-xt-relationalStore-bigquery-execution-tests/src/test/java/org/finos/legend/engine/plan/execution/stores/relational/connection/test/TestBigQueryWithGCPApplicationDefaultCredentialsFlow.java
@@ -12,8 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.engine.authentication.flows;
+package org.finos.legend.engine.plan.execution.stores.relational.connection.test;
 
+import org.finos.legend.engine.authentication.flows.BigQueryWithGCPApplicationDefaultCredentialsFlow;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.authentication.GCPApplicationDefaultCredentialsAuthenticationStrategy;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.specification.BigQueryDatasourceSpecification;
 import org.finos.legend.engine.shared.core.identity.Credential;
